harmonic centrality nearest neighbor calculation that uses the harmonic centrality of the highest neighbors node as a seed node proxy for the centrality of that node (with a discount factor)

+++ crates/core/src/live_index/crawler/budgets.rs
@@ -143,7 +143,7 @@ impl SiteBudgets {
         if budget == 0 {
             Some(Duration::from_millis(MILLIS_PER_DAY)) // once per day
         } else {
-            Some(Duration::from_millis(MILLIS_PER_DAY / budget as u64))
+            Some(Duration::from_millis(MILLIS_PER_DAY / budget))
         }
     }
 }
